+++ modules/RpcApi/RpcApiRequestHandler.hpp
@@ -20,8 +20,8 @@ public:
     types::json_rpc_api::ErrorResObj setACCharging(const int32_t evse_index, bool charging_allowed, bool max_current, std::optional<int> phase_count) override;
     types::json_rpc_api::ErrorResObj setACChargingCurrent(const int32_t evse_index, float max_current) override;
     types::json_rpc_api::ErrorResObj setACChargingPhaseCount(const int32_t evse_index, int phase_count) override;
-    types::json_rpc_api::ErrorResObj setDCCharging(const int32_t evse_index, bool charging_allowed, int max_power) override;
-    types::json_rpc_api::ErrorResObj setDCChargingPower(const int32_t evse_index, int max_power) override;
+    types::json_rpc_api::ErrorResObj setDCCharging(const int32_t evse_index, bool charging_allowed, float max_power) override;
+    types::json_rpc_api::ErrorResObj setDCChargingPower(const int32_t evse_index, float max_power) override;
     types::json_rpc_api::ErrorResObj enableConnector(const int32_t evse_index, int connector_id, bool enable, int priority) override;
 private:
     // Add any private member variables or methods here
